Changed SchemaRegistryList to SchemaList
